Add column seek in Peak+Tile Matrix; RegionSelect
Implemented tests for column seeking and uncovered a couple bugs to fix.
I wouldn't recommend using column seeking on Peak/Tile matrix for
performance reasons, but it's good to have available in a pinch.

RegionSelect is not yet implemented in R, but looks to be fine based on
C++ tests