Implement MergeFragments utility
Merging multiple fragments -- the only tricky bit conceptually is how to
handle merging chromosomes appropriately. The solution I settled on for
now is to require identical chromosome levels, and for each iterator
to cover the chromosomes in order. This should be no problem for
StoredFragments but might be tough for BedFragments.

The tests here were a little tricky to implement, since the ordering of
fragments with identical starts is arbitrary. But I was able to make a
setup that gets a reproducible order for testing purposes.