Update MatrixWriter and FragmentWriter interfaces
Made write take a *Loader object instead of a *Iterator.
This sometimes requires an extra line for the callee, while saving the caller
one line and a variable declaration.

Also made a quick fix on the StoredFragmentsWriter so that it can accept
loaders that load in larger blocks than its array storage can write.

This is tested in the fragmentIO tests, which have also been updated to
avoid needing external data files to run