Added InsertionIterator
InsertionIterator is a convenience class to merge fragments
(start + ends) into just insertion coordinates. It appears to be
functional when tested, and I expect good performance though
nothing is certain in this world :)

The only caveat currently is that seeking may not work quite as expected
in that after a seek it will output some susbet of the insertions prior
to the seeked location. But after reaching the seeked location all
insertion will be present.