Implemented Matrix multiplies for Shift transform
This is an optimization that takes advantage of mathematical
rearranging we can do on the shift transform.

For the future we could
implement the row/column stats as well, but I'm not sure about how to
handle the non-zero entry counts yet. (Ultimately, the best idea is to
make it possible to compute mean+variance without also returning the
non-zero entry counts)