Add matrix concatenation, update matrix XPtr handling
This update to XPtr handling mirrors the changes made already for
fragments. The idea is that R will keep a list of XPtr objects, so the
C++ code is guaranteed none of the dependent objects will disappear
out from it due to an unlucky GC-timing.
This also has the fortunate side-effect of letting us perform PCA without
needing to reconstruct the XPtr iterator for each mat-vec product