C++ matrix row select bugfix
Cause of this issue was capacity() advertising a lot, but load()
advancing by a little, causing the same data to be read twice.

For the future, either I should consider switching loading interface to
caller-allocated memory, as I've run into other bugs with seeking.
This might also alleviate the nonsense of parceling out subsets of
data in an attempt to avoid naive callers iterating data in
larger-than-cache chunks.