Support row-major saving of matrix files
This overhauls the matrix disk format to allow marking matrices as
row-major. All internal calculations are still done col-major, meaning
that row-major matrices are logically transposed during read + write.

This changes some of the names of the matrix data arrays to match
scipy/10x/AnnData conventions and so the names make sense also for
row-major matrices.

The purpose of this change is to remove footguns when saving transposed
matrices. Now the expensive changing of physical storage layout only
happens with `transpose_storage_order`, and in day-to-day use users
don't need to worry about the transpose status of their matrices before
and after loading.