Tile-based peak calling
Adds peak calling based on an overlapping tiling of the genome.
This utilizes the new `merge_cells` functionality. On a 3k PBMC dataset,
I observed a drop in total time from 80 seconds to 48 seconds when doing
the merge rather than matrix multiply. (This is for 200bp tiles with 3
different phasings)

The major point of difference compared to ArchR is not doing sample
splitting