Fix bug with global scale on transposed matrix
The logic for iterate_matrix on a transposed TransformScaleShift matrix
made a row_params of incorrect dimensions when only a global scale or
global shift was applied.

This bug has been fixed, and additional checking has been added to
MatrixTransform.cpp so that it will have a better error message in the
future.