Merge fragments: support non-matching chr sets
This allows merging fragments with non-matching chromosome names.
Chromosmes are matched by name, and the resulting merged object has a
union of the input chromosome names.

The implementation is also simplified by adding a requirement that all
the inputs be seekable, rather than allowing non-seekable inputs with
known chromosome names.