ConcatCols bugfix
This bug seemed to trigger when calling nextCol() after a seek().
nextCol() assumed that after finishing one matrix, the next matrix would
be just before column 0, but this is not necessarily true when a seek
has occurred. This fixes it, and the upcoming `all_matrix_inputs()`
tests would trigger the bug if it were re-introduced.