Add C++ SNN graph construction
This addds a C++ implementation to construct the SNN graph. Previously,
the dgCMatrix version would crash on large datasets. Initial C++
attempts suffered from large memory usage, which this current approach
does not fully alleviate but makes better (<16GB RAM for 1 million cells
with 20 nearest neighbors)

Useful numbers: 10x 1 million cell dataset with 20 KNNs. There are about
500 million shared nearest neighbor instances, from 340 million unique
cell-cell pairs. Trimming down to the Seurat defaults of >2 shared
neighbors gets us down to 35 million unique. Because the trimming is
a default and saves a bunch of space, we do it in C++.

The data structures used are pretty basic, but seem to do the job okay.
I did try using std::unordered_map but that was slow and ran out of
memory faster.