[r] Fix perf subsetting large conatenated matrices
The logic for subsetting rbind or cbind matrices was taking time
O(length(selection)* # sub-matrices), which could be quite slow for large
datasets. Now, rather than performing a linear search through the
selection indices to find the indices in range for each sub-matrix,
we just do a binary search which eliminates the performance issue.