[cpp] Improve stats calculation in row/col reorder
Motivation: The subset assignment operator results in a concatenation
layer followed by a re-ordering of the rows/cols. We don't want this
outer re-order layer to prevent any parallelism or other optimizations
available on the inner layers, so it's time to have all the stats
specialized for row/col re-ordering.